backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
proxmoxer==2.0.1
requests==2.31.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
boto3==1.34.100
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.2.1
